+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Development
 
 - Update wetterdienst explorer with clickable stations and slighly changed layout
 - Improve radar tests and certain dict comparisons
+- Fix problem with numeric column names in method gain_of_value_pairs
 
 0.50.0 (03.12.2022)
 *******************